requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
alabaster=0.7.12=pypi_0
babel=2.10.1=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.9.24=py38h06a4308_0
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.12=pypi_0
cryptography=37.0.1=py38h9ce1e76_0
cuda-python=11.6.0=py38h3fd9d12_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
docutils=0.17.1=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
einops=0.6.1=pypi_0
et-xmlfile=1.1.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.12.2=pypi_0
fonttools=4.31.2=pypi_0
freetype=2.11.0=h70c0345_0
fsspec=2023.6.0=pypi_0
gdown=4.7.1=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.16.4=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.31.1=pypi_0
imagesize=1.3.0=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jinja2=3.1.1=pypi_0
joblib=1.3.2=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
munch=4.0.0=pypi_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
numpy=1.21.5=py38he7a7128_2
numpy-base=1.21.5=py38hf524024_2
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.1=pypi_0
pillow=9.0.1=pypi_0
pip=21.2.4=py38h06a4308_0
plotly=5.11.0=pypi_0
pretrainedmodels=0.7.4=pypi_0
progress-bar=8=pypi_0
protobuf=4.24.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch-mutex=1.0=cpu
pytorchcv=0.0.67=pypi_0
pytz=2022.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
safetensors=0.3.2=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.8.0=pypi_0
segmentation-models-pytorch=0.3.3=pypi_0
setuptools=58.0.4=py38h06a4308_0
shapely=2.0.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
snowballstemmer=2.2.0=pypi_0
sphinx=4.5.0=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=2.0.0=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-katex=0.8.6=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.5=pypi_0
sqlite=3.38.0=hc218d9a_0
tenacity=8.1.0=pypi_0
tensorboardx=2.6.2=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.9.2=pypi_0
tk=8.6.11=h1ccaba5_0
torch=1.11.0+cu113=pypi_0
torch-dct=0.1.5=pypi_0
torchaudio=0.11.0=py38_cpu
torchvision=0.10.0=pypi_0
ttach=0.0.3=pypi_0
typing_extensions=4.3.0=py38h06a4308_0
urllib3=1.26.9=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xlrd=1.2.0=pypi_0
xz=5.2.5=h7b6447c_0
zipp=3.8.0=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
